+++ src/index.js
@@ -461,10 +461,20 @@ export default class Carousel extends React.Component {
 
   autoplayIterator() {
     if (this.props.wrapAround) {
-      this.nextSlide();
+      if (this.props.autoplayReverse) {
+        this.previousSlide();
+      } else {
+        this.nextSlide();
+      }
       return;
     }
-    if (
+    if (this.props.autoplayReverse) {
+      if (this.state.currentSlide !== 0) {
+        this.previousSlide();
+      } else {
+        this.stopAutoplay();
+      }
+    } else if (
       this.state.currentSlide !==
       this.state.slideCount - this.state.slidesToShow
     ) {
@@ -872,7 +882,9 @@ export default class Carousel extends React.Component {
     } = this.props;
     const duration =
       this.state.dragging ||
-      this.state.resetWrapAroundPosition ||
+      (!this.state.dragging &&
+        this.state.resetWrapAroundPosition &&
+        this.props.wrapAround) ||
       disableAnimation ||
       !this.state.hasInteraction
         ? 0
@@ -887,12 +899,7 @@ export default class Carousel extends React.Component {
     const touchEvents = this.getTouchEvents();
     const mouseEvents = this.getMouseEvents();
     const TransitionControl = Transitions[this.props.transitionMode];
-    const validChildren = addAccessibility(
-      getValidChildren(this.props.children),
-      slidesToShow,
-      currentSlide
-    );
-    const transitionProps = getTransitionProps(this.props, this.state);
+    const validChildren = getValidChildren(this.props.children);
 
     return (
       <div
@@ -961,8 +968,12 @@ export default class Carousel extends React.Component {
               };
             }}
             children={({ tx, ty }) => (
-              <TransitionControl {...transitionProps} deltaX={tx} deltaY={ty}>
-                {validChildren}
+              <TransitionControl
+                {...getTransitionProps(this.props, this.state)}
+                deltaX={tx}
+                deltaY={ty}
+              >
+                {addAccessibility(validChildren, slidesToShow, currentSlide)}
               </TransitionControl>
             )}
           />
@@ -987,6 +998,7 @@ Carousel.propTypes = {
   autoGenerateStyleTag: PropTypes.bool,
   autoplay: PropTypes.bool,
   autoplayInterval: PropTypes.number,
+  autoplayReverse: PropTypes.bool,
   beforeSlide: PropTypes.func,
   cellAlign: PropTypes.oneOf(['left', 'center', 'right']),
   cellSpacing: PropTypes.number,
@@ -1036,6 +1048,7 @@ Carousel.defaultProps = {
   autoGenerateStyleTag: true,
   autoplay: false,
   autoplayInterval: 3000,
+  autoplayReverse: false,
   beforeSlide() {},
   cellAlign: 'left',
   cellSpacing: 0,